+++ src/main/java/pl/thedeem/intellij/dql/fileProviders/DQLRecordVirtualFile.java
@@ -2,7 +2,7 @@ package pl.thedeem.intellij.dql.fileProviders;
 
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
-import pl.thedeem.intellij.dql.components.execution.DQLRecordPanel;
+import pl.thedeem.intellij.dql.exec.panel.DQLRecordPanel;
 
 import javax.swing.*;
 import java.util.Map;